refactor: update project structure and clean up configurations
- Removed app.config.ts and AdminNav.vue components to streamline the project.
- Updated .gitignore to exclude the dist directory.
- Adjusted import paths in various components and pages for consistency.
- Cleaned up tsconfig.json to reference specific Nuxt TypeScript configurations.
- Enhanced comments in nuxt.config.ts for clarity on ISR settings.
