--- app/components/common/AdminNav.vue
@@ -1,27 +1,0 @@
-<template>
-  <div v-if="auth.authenticated" class="admin-nav">
-    <a 
-      href="http://localhost:1337/admin" 
-      target="_blank"
-      class="flex items-center gap-2 px-4 py-2 bg-primary hover:opacity-90 dark:bg-primary-400 dark:hover:opacity-90 text-white rounded-md transition"
-    >
-      <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
-        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
-      </svg>
-      Strapi Admin
-    </a>
-  </div>
-</template>
-
-<script setup lang="ts">
-import { useAuth } from '~/composables/useAuth';
-
-const auth = useAuth();
-</script>
-
-<style scoped>
-.admin-nav {
-  margin-left: 1rem;
-}
-</style>
