+++ tsconfig.json
@@ -1,25 +1,17 @@
 {
-  // https://nuxt.com/docs/guide/concepts/typescript
-  "extends": "./.nuxt/tsconfig.json",
-  "compilerOptions": {
-    "types": ["nuxt", "node"],
-    "strict": true,
-    "noImplicitAny": true,
-    "noImplicitOverride": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true,
-    "forceConsistentCasingInFileNames": true,
-    "exactOptionalPropertyTypes": false,
-    "useUnknownInCatchVariables": true,
-    "noUncheckedIndexedAccess": true,
-    "skipLibCheck": true,
-    "baseUrl": ".",
-    "paths": {
-      "~/*": ["./*"],
-      "@/*": ["./*"]
+  "files": [],
+  "references": [
+    {
+      "path": "./.nuxt/tsconfig.app.json"
+    },
+    {
+      "path": "./.nuxt/tsconfig.server.json"
+    },
+    {
+      "path": "./.nuxt/tsconfig.shared.json"
+    },
+    {
+      "path": "./.nuxt/tsconfig.node.json"
     }
-  },
-  "include": [".nuxt/nuxt.d.ts", "**/*"],
-  "exclude": ["../node_modules/@nuxt/image/**", "../../node_modules/@nuxt/image/**"]
+  ]
 }
